include ctype.h for isdigit()

+++ tests/server/sws.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: sws.c,v 1.71 2005-05-18 20:01:01 bagder Exp $
+ * $Id: sws.c,v 1.72 2005-05-19 09:55:53 bagder Exp $
  ***************************************************************************/
 
 /* sws.c: simple (silly?) web server
@@ -37,6 +37,7 @@
 #include <time.h>
 #include <sys/time.h>
 #include <sys/types.h>
+#include <ctype.h>
 
 #ifdef HAVE_UNISTD_H
 #include <unistd.h>
@@ -228,7 +229,7 @@ int ProcessRequest(struct httprequest *req)
       ptr++; /* skip the slash */
 
       /* skip all non-numericals following the slash */
-      while(*ptr && !isdigit(*ptr))
+      while(*ptr && !isdigit((int)*ptr))
         ptr++;
 
       req->testno = strtol(ptr, &ptr, 10);
